Revert "[BE022] payment process"

+++ src/main/java/com/tttm/birdfarmshop/Utils/Response/OrderResponse.java
@@ -28,7 +28,7 @@ public class OrderResponse {
 
     private String note;
 
-    private Integer orderAmount;
+    private Float orderAmount;
 
     private Date orderDate;
